Move LightningDistanceField class to separate file
As per our code style, we should have one class per file.
Tiny helper structs like GroundingLocation can be forgiven for the purpose of readability though.

Contributes to issue CURA-8550.
