+++ CMakeLists.txt
@@ -179,6 +179,7 @@ set(engine_SRCS # Except main.cpp.
     src/infill/ImageBasedDensityProvider.cpp
     src/infill/NoZigZagConnectorProcessor.cpp
     src/infill/ZigzagConnectorProcessor.cpp
+    src/infill/LightningDistanceField.cpp
     src/infill/LightningGenerator.cpp
     src/infill/LightningLayer.cpp
     src/infill/LightningTreeNode.cpp
